+++ src/persistence/get_connection.py
@@ -1,0 +1,6 @@
+import pymysql
+
+def get_connection():
+    return pymysql.connect(host="localhost", port=33066, user="root", password="password", database="brewapp")
+
+connection = get_connection()
